+++ src/blog/Models/BlogPost.cs
@@ -7,8 +7,8 @@ namespace blog.Models
 {
   public class BlogPost
   {
-    public int Id { get; set; }
-    public int? AuthorId { get; set; }
+    public long Id { get; set; }
+    public long? AuthorId { get; set; }
     public Author Author { get; set; }
     public string Title { get; set; }
     public string Description { get; set; }
